Extract TestService helper to shared module
I've used this helper enough times that I think it's OK to extact it
to its own module.
