+++ test/e2e/app-dir/ppr-navigations/ppr-navigations.test.ts
@@ -1,6 +1,7 @@
 import { createNext } from 'e2e-utils'
 import { findPort } from 'next-test-utils'
-import http from 'http'
+import { createTestDataServer } from 'test-data-service/writer'
+import { createTestLog } from 'test-log'
 
 describe('ppr-navigations', () => {
   if ((global as any).isNextDev) {
@@ -198,238 +199,3 @@ describe('ppr-navigations', () => {
     }
   )
 })
-
-// NOTE: I've intentionally not yet moved these helpers into a separate
-// module, to avoid early abstraction. I will if/when we start using them for
-// other tests. They are based on the testing patterns we use all over the React
-// codebase, so I'm reasonably confident in them.
-
-type TestDataResponse = {
-  _res: http.ServerResponse
-  resolve: (value?: string) => any
-  reject: (value: any) => any
-}
-
-type TestDataServer = {
-  _server: http.Server
-  listen: (port: number) => void
-  close: () => void
-}
-
-// Creates a lightweight HTTP server for use in e2e testing. This simulates the
-// data service that would be used in a real Next.js application, whether it's
-// direct database access, an ORM, or a higher-level data access layer. The e2e
-// test can observe when individual requests are received, and control the
-// timing of when the data is fulfilled, without needing to mock any lower
-// level I/O.
-//
-// Receives requests of the form: /?key=foo
-//
-// Responds in plain text. By default, the response is the key itself, but the
-// e2e test can respond with any text it wants.
-//
-// Examples:
-//   response.resolve() // Responds with the key itself
-//   response.resolve('custom') // Responds with custom text
-//   response.reject(new Error('oops!')) // Responds with a 500 error
-//
-// Based on the AsyncText pattern used in the React repo.
-function createTestDataServer(
-  onRequest: (key: string, response: TestDataResponse) => any
-): TestDataServer {
-  const httpServer = http.createServer(async (req, res) => {
-    const searchParams = new URL(req.url, 'http://n').searchParams
-    const key = searchParams.get('key')
-
-    if (typeof key !== 'string') {
-      res.statusCode = 400
-      const msg = 'Missing key parameter'
-      res.end(msg)
-      return
-    }
-
-    const response: TestDataResponse = {
-      _res: res,
-      resolve(value?: string) {
-        res.end(value === undefined ? key : value)
-      },
-      reject(error: Error, status?: number) {
-        res.statusCode = status ?? 500
-        res.end(error.message ?? `Failed to fetch data for "${key}"`)
-      },
-    }
-
-    try {
-      const result = await onRequest(key, response)
-      if (typeof result === 'string') {
-        response.resolve(result)
-      }
-    } catch (error) {
-      response.reject(error)
-    }
-  })
-  return {
-    _server: httpServer,
-    listen(port: number) {
-      httpServer.listen(port)
-    },
-    close() {
-      httpServer.close()
-    },
-  }
-}
-
-// Creates an event log. You can write to this during testing and then assert
-// on the result.
-//
-// The main use case is for asynchronous e2e tests. It provides a `waitFor`
-// method that resolves when the log matches some expected asynchronous sequence
-// of events. This is an alternative to setting up a timer loop. It helps catch
-// subtle mistakes where the order of events is not expected, or the same
-// event happens more than it should.
-//
-// Based on the Scheduler.log pattern used in the React repo.
-function createTestLog() {
-  let events = []
-
-  // Represents a pending waitFor call.
-  let pendingExpectation: null | {
-    resolve: () => void
-    reject: (error: Error) => void
-    expectedEvents: Array<any>
-    error: Error
-  } = null
-
-  function log(value: any) {
-    // Add to the event log.
-    events.push(value)
-
-    // Check if we've reached the end of the expected log. If there's a
-    // pending waitFor, and we've reached the last of the expected events, this
-    // will resolve the promise.
-    pingExpectation()
-  }
-
-  function assert(expectedEvents: any[]) {
-    if (pendingExpectation !== null) {
-      const error = new Error('Cannot assert while a waitFor() is pending.')
-      Error.captureStackTrace(error, assert)
-      throw error
-    }
-
-    const actualEvents = events
-    events = []
-
-    if (!areLogsEqual(expectedEvents, actualEvents)) {
-      // Capture the stack trace of `assert` so that Jest will report the
-      // error as originating from the `assert` call instead of here.
-      const error = new Error(
-        'Expected sequence of events did not occur.\n\n' +
-          createDiff(expectedEvents, actualEvents)
-      )
-      Error.captureStackTrace(error, assert)
-      throw error
-    }
-  }
-
-  function waitFor(expectedEvents: any[], timeout: number = 5000) {
-    // Returns a promise that resolves when the event log matches the
-    // expected sequence.
-
-    // Capture the stack trace of `waitFor` so that if an inner assertion fails,
-    // Jest will report the error as originating from the `waitFor` call instead
-    // of inside this module's implementation.
-    const error = new Error()
-    Error.captureStackTrace(error, waitFor)
-
-    if (pendingExpectation !== null) {
-      error.message = 'A previous waitFor() is still pending.'
-      throw error
-    }
-
-    let resolve
-    let reject
-    const promise = new Promise<void>((res, rej) => {
-      resolve = res
-      reject = rej
-    })
-
-    const thisExpectation = {
-      resolve,
-      reject,
-      expectedEvents,
-      error,
-    }
-    pendingExpectation = thisExpectation
-
-    setTimeout(() => {
-      if (pendingExpectation === thisExpectation) {
-        error.message = `waitFor timed out after ${timeout}ms`
-        reject(error)
-      }
-    }, timeout)
-
-    pingExpectation()
-
-    return promise
-  }
-
-  function pingExpectation() {
-    if (pendingExpectation !== null) {
-      const expectedEvents = pendingExpectation.expectedEvents
-      if (events.length < expectedEvents.length) {
-        return
-      }
-
-      if (areLogsEqual(expectedEvents, events)) {
-        // We've reached the end of the expected log. Resolve the promise and
-        // reset the log.
-        events = []
-        pendingExpectation.resolve()
-        pendingExpectation = null
-      } else {
-        // The log does not match what was expected by the test. Reject the
-        // promise and reset the log.
-
-        // Use the error object that we captured at the start of the `waitFor`
-        // call. Jest will show that the error originated from `waitFor` call
-        // instead of inside this internal function.
-        const error = pendingExpectation.error
-        error.message =
-          'Expected sequence of events did not occur.\n\n' +
-          createDiff(expectedEvents, events)
-
-        events = []
-        pendingExpectation.reject(error)
-        pendingExpectation = null
-      }
-    }
-  }
-
-  function createDiff(expected, actual) {
-    // TODO: Jest exposes the diffing utility that it uses for `expect`.
-    // We could use that here for nicer output.
-    return `
-Expected: ${JSON.stringify(expected)}
-Actual:   ${JSON.stringify(actual)}
-`
-  }
-
-  function areLogsEqual(a, b) {
-    if (a.length !== b.length) {
-      return false
-    }
-    for (let i = 0; i < a.length; i++) {
-      if (a[i] !== b[i]) {
-        return false
-      }
-    }
-    return true
-  }
-
-  return {
-    log,
-    waitFor,
-    assert,
-  }
-}
