+++ tsconfig.base.json
@@ -16,7 +16,9 @@
       "next-test-utils": ["test/lib/next-test-utils"],
       "amp-test-utils": ["test/lib/amp-test-utils"],
       "next-webdriver": ["test/lib/next-webdriver"],
-      "e2e-utils": ["test/lib/e2e-utils"]
+      "e2e-utils": ["test/lib/e2e-utils"],
+      "test-data-service/*": ["test/lib/test-data-service/*"],
+      "test-log": ["test/lib/test-log"]
     }
   }
 }
